+++ types/index.d.ts
@@ -1,2 +1,2 @@
 export * from './azure-key-vault';
-export { default as AzureKeyVault } from './azure-key-vault.service';
+export * from './azure-key-vault.service';