+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [1.1.5] - 2021-05-01
 -   types fix.
+-   delete and restore fix.
 
 ## [1.1.3] - 2021-04-30
 -   fixed bad nesting handling.